Move createTRBRequest() utility function for store tests into store_test.go, use universally
Remove now-redundant createTRBRequestWithContext() function
